Fix progress member visiblity
