+++ src/epg_station_api/model.rs
@@ -120,3 +120,18 @@ impl RecordedQuery {
         parameters
     }
 }
+
+pub struct TransferProgress {
+    pub(in crate::epg_station_api) current_bytes: u64,
+    pub(in crate::epg_station_api) total_bytes: u64,
+}
+
+impl TransferProgress {
+    pub fn current_bytes(&self) -> u64 {
+        self.current_bytes
+    }
+
+    pub fn total_bytes(&self) -> u64 {
+        self.total_bytes
+    }
+}
